requirements (1).txt
streamlit==1.33.0
tensorflow-cpu==2.15.0
ultralytics==8.1.0
Pillow==9.5.0
numpy==1.24.3
opencv-python-headless==4.8.1.78
